The crossbar service only mapping the config file to the container, not the whole directory (see docker-compose.yml).
The running crossbar router used the working directory to store the running PID file and some certificates. If it has stopped improperly, the files remained in the host filesystem -> next time the crossbar executable was not able to started, because of the PID file it has mean recognized incorrectly as a running service.